big changes :), in reducers and action and connect whith backend

+++ api/app/controllers/links.ts
@@ -6,23 +6,18 @@ export const createLink=async(req:Req, res:Res, next: Next)=>{
 
     const {nameLink,url,courseId}=req.body;
 
-
     if(!nameLink || !url || ! courseId) return res.send({error:'Missing fields',content:null});
 
     try{
 
-        const newLink=prisma.link.create({
-
+        const newLink= await prisma.link.create({
             data:{
                 name:nameLink,
                 url:url,
                 courseId
-            },
-            select:{
-                name:true,
-                url:true
             }
         });
+        console.log('el nuevo link es: ',newLink)
 
         if(!newLink) return res.send({error:'newLink Not found',content:null});
         return res.send({error:null,content:newLink})
@@ -44,6 +39,15 @@ export const deleteLink=async(req: Req, res: Res,next: Next)=>{
     if(!nameLink) return res.send({error:'nameLink not exist',content:null});
 
     try{
+        const linkExist=await prisma.link.findUnique({
+            where:{
+                name:nameLink
+            }
+        });
+
+        if(!linkExist) return res.status(404).send({error:'link not exist',content:null});
+
+
         const oldLink=await prisma.link.delete({
             where:{
                 name:nameLink
@@ -62,6 +66,5 @@ export const deleteLink=async(req: Req, res: Res,next: Next)=>{
         console.error(error);
         return res.send({error, content:null})
     }
-
 
 }
